add if ($task->user_id == \Auth::id()

+++ app/Http/Controllers/TasksController.php
@@ -30,13 +30,6 @@ class TasksController extends Controller
         // Welcomeビューでそれらを表示
         return view('welcome', $data);
         
-        // タスク一覧を取得
-        $tasks = Task::all();
-
-        // タスク一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
     }
 
     /**
@@ -61,7 +54,7 @@ class TasksController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-        // postでtasks/にアクセスされた場合の「新規登録処理」
+    // postでtasks/にアクセスされた場合の「新規登録処理」
     public function store(Request $request)
     {
         // バリデーション
@@ -100,9 +93,9 @@ class TasksController extends Controller
         //taskの所有者とログインユーザーが一致しているかどうか
         //一致していなければトップページへリダイレクトさせる
         // タスク詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
-        ]); }
+            return view('tasks.show', [
+                'task' => $task,
+            ]); }
         
         else{
            return redirect('/'); 
@@ -123,9 +116,9 @@ class TasksController extends Controller
         if ($task->user_id == \Auth::id()) {
 
         // タスク編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-        ]); }
+            return view('tasks.edit', [
+                'task' => $task,
+            ]); }
         
          else{
            return redirect('/'); 
@@ -150,10 +143,13 @@ class TasksController extends Controller
 
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
+        
+        if ($task->user_id == \Auth::id()) {
         // タスクを更新
-        $task->status = $request->status;    // 追加
-        $task->content = $request->content;
-        $task->save();
+            $task->status = $request->status;    // 追加
+            $task->content = $request->content;
+            $task->save();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
@@ -170,9 +166,10 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
-
+        if ($task->user_id == \Auth::id()) {
+            // タスクを削除
+            $task->delete();
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
